Implemented modification of user info and account info

+++ src/main/java/com/revature/repository/AccountDAOIMPL.java
@@ -133,26 +133,18 @@ public class AccountDAOIMPL implements AccountDAO {
 		try (Connection conn = ConnectionUtil.getConnection()) {
 			
 			
-			String sql = "UPDATE accounts SET"
-			+ "account_id = ?,"
-			+ " account_type = ?,"
-			+ " account_number = ?,"
-			+ " balance = ?,"
-			+ " interest_rate = ?,"
-			+ " is_joint = ?,"
-			+ "pin_number = ?"
-			+ "WHERE account_id = ?;"; 
+			String sql = "UPDATE public.accounts SET account_type = ?, account_number = ?, balance = ?, interest_rate = ?, is_joint = ?, pin_number = ? WHERE account_id = ?;"; 
 					
 			
 			PreparedStatement stm = conn.prepareStatement(sql);
-			stm.setInt(1, id);
-			stm.setString(2, type);
-			stm.setInt(3, number);
-			stm.setDouble(4, balance);
-			stm.setDouble(5, interest);
-			stm.setBoolean(6, isJoint);
-			stm.setInt(7, pin);
-			stm.setInt(8, id);
+			
+			stm.setString(1, type);
+			stm.setInt(2, number);
+			stm.setDouble(3, balance);
+			stm.setDouble(4, interest);
+			stm.setBoolean(5, isJoint);
+			stm.setInt(6, pin);
+			stm.setInt(7, id);
 			if(!stm.execute()) {
 				return false;
 			}
@@ -168,8 +160,7 @@ public class AccountDAOIMPL implements AccountDAO {
 		int id = a.getId();
 		try (Connection conn = ConnectionUtil.getConnection()) {
 						
-					String sql = "DELETE FROM accounts"
-					+ "WHERE account_id = ? ;"; 
+					String sql = "DELETE FROM public.accounts WHERE account_id = ?;"; 
 					
 					PreparedStatement stm = conn.prepareStatement(sql);
 					stm.setInt(1, id);
@@ -191,7 +182,7 @@ public class AccountDAOIMPL implements AccountDAO {
 		
 		try (Connection con = ConnectionUtil.getConnection()) {
 				
-			String sql = "SELECT * FROM accounts WHERE pin_number = ? ;";
+			String sql = "SELECT * FROM public.accounts WHERE pin_number = ?;";
 			
 			PreparedStatement stmt = con.prepareStatement(sql);
 			stmt.setInt(1, pinNumber);
@@ -230,26 +221,17 @@ public class AccountDAOIMPL implements AccountDAO {
 		try (Connection conn = ConnectionUtil.getConnection()) {
 			
 			
-			String sql = "UPDATE accounts SET "
-			+ "account_id = ?,"
-			+ "account_type = ?,"
-			+ "account_number = ?,"
-			+ "balance = ?,"
-			+ "interest_rate = ?,"
-			+ "is_joint = ?,"
-			+ "pin_number = ?"
-			+ "WHERE account_id = ?;"; 
+			String sql = "UPDATE public.accounts SET account_type = ?, account_number = ?, balance = ?, interest_rate = ?, is_joint = ?, pin_number = ? WHERE account_id = ?;"; 
 					
 			
 			PreparedStatement stm = conn.prepareStatement(sql);
-			stm.setInt(1, id);
-			stm.setString(2, type);
-			stm.setInt(3, number);
-			stm.setDouble(4, balance);
-			stm.setDouble(5, interest);
-			stm.setBoolean(6, isJoint);
-			stm.setInt(7, pin);
-			stm.setInt(8, id);
+			stm.setString(1, type);
+			stm.setInt(2, number);
+			stm.setDouble(3, balance);
+			stm.setDouble(4, interest);
+			stm.setBoolean(5, isJoint);
+			stm.setInt(6, pin);
+			stm.setInt(7, id);
 			if(!stm.execute()) {
 				return false;
 			}
